moved to interface based approach
