+++ packages/dev/core/src/Culling/index.ts
@@ -2,6 +2,7 @@
 export * from "./boundingBox";
 export * from "./boundingInfo";
 export * from "./Helper/boundingInfoHelper";
+export * from "./Helper/transformFeedbackBoundingHelper";
 export * from "./boundingSphere";
 export * from "./Octrees/index";
 export * from "./ray";
